# .api/apis/eupago/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "REST API - ApiKey Auth",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://sandbox.eupago.pt/api/"
    }
  ],
  "components": {
    "securitySchemes": {
      "sec0": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "x-bearer-format": "",
        "x-default": ""
      }
    }
  },
  "security": [
    {
      "sec0": []
    }
  ],
  "paths": {
...
    "/v1.02/paybylink/create": {
      "post": {
        "summary": "PayByLink",
        "description": "Creates a request to allow the customer to pay using PayByLink.",
        "operationId": "paybylink",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "payment"
                ],
                "properties": {
                  "payment": {
                    "type": "object",
                    "description": "Informations regarding the payment",
                    "properties": {
                      "identifier": {
                        "type": "string"
                      },
                      "amount": {
                        "type": "object",
                        "required": [
                          "value",
                          "currency"
                        ],
                        "properties": {
                          "value": {
                            "type": "number",
                            "description": "Payment amount",
                            "format": "float"
                          },
                          "currency": {
                            "type": "string",
                            "default": "EUR"
                          }
                        }
                      },
                      "shipping": {
                        "type": "object",
                        "required": [
                          "value"
                        ],
                        "properties": {
                          "value": {
                            "type": "number",
                            "description": "Shipping amount",
                            "default": 0,
                            "format": "float"
                          },
                          "currency": {
                            "type": "string",
                            "default": "EUR"
                          }
                        }
                      },
                      "lang": {
                        "type": "string",
                        "description": "Defines the Language of the email notifications.",
                        "default": "PT"
                      },
                      "expirationDate": {
                        "type": "string",
                        "description": "Defines the date when the link will expire (Default value is 24 hours after the link creation)",
                        "default": "2099-12-31 23:59:59",
                        "format": "date"
                      },
                      "successUrl": {
                        "type": "string",
                        "description": "The payer will be forwarded to this url if the payment process is successful."
                      },
                      "failUrl": {
                        "type": "string",
                        "description": "The payer will be forwarded to this url if the payment process fails."
                      },
                      "backUrl": {
                        "type": "string",
                        "description": "The payer will be forwarded to this url if the user presses back."
                      }
                    }
                  },
                  "products": {
                    "type": "array",
                    "description": "All products in the order. This param may include different objects (products).",
                    "items": {
                      "properties": {
                        "sku": {
                          "type": "string",
                          "description": "Product code (Required to pay with Cofidis Pay)"
                        },
                        "name": {
                          "type": "string",
                          "description": "Product name"
                        },
                        "value": {
                          "type": "number",
                          "description": "Product price",
                          "format": "float"
                        },
                        "quantity": {
                          "type": "integer",
                          "description": "Number of times this product is ordered  (Required to pay with Cofidis Pay)",
                          "format": "int32"
                        },
                        "tax": {
                          "type": "number",
                          "description": "Tax percentage (Required to pay with Cofidis Pay)",
                          "format": "float"
                        },
                        "discount": {
                          "type": "number",
                          "description": "(Required to pay with Cofidis Pay)",
                          "format": "float"
                        },
                        "description": {
                          "type": "string",
                          "description": "Text describing the product (Required to pay with Cofidis Pay)"
                        }
                      },
                      "required": [
                        "value"
                      ],
                      "type": "object"
                    }
                  },
                  "customer": {
                    "type": "object",
                    "description": "Informations regarding customer's notifications",
                    "properties": {
                      "notify": {
                        "type": "boolean"
                      },
                      "email": {
                        "type": "string",
                        "description": "Customer's email"
                      },
                      "nome": {
                        "type": "string",
                        "description": "Customer's name"
                      }
                    }
                  }
                }
              },
              "examples": {
                "Example": {
                  "value": {
                    "payment": {
                      "identifier": "New Test",
                      "amount": {
                        "value": 100.3,
                        "currency": "EUR"
                      },
                      "shipping": {
                        "value": 20.1,
                        "currency": "EUR"
                      },
                      "lang": "PT"
                    },
                    "products": [
                      {
                        "sku": "111111111",
                        "name": "Product 1",
                        "value": 50.1,
                        "quantity": 1,
                        "tax": 23,
                        "discount": 0,
                        "description": "Description of product 1"
                      },
                      {
                        "sku": "111111111",
                        "name": "Product 2",
                        "value": 30.1,
                        "quantity": 1,
                        "tax": 23,
                        "discount": 0,
                        "description": "Description of product 2"
                      }
                    ],
                    "customer": {
                      "notify": true,
                      "email": "exemplo@eupago.pt",
                      "nome": "Nome"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "200",
            "content": {
              "application/json": {
                "examples": {
                  "OK": {
                    "value": "{\n    \"transactionStatus\": \"Success\",\n    \"transactionID\": \"af3df607c6724870be962a69cac30b99\",\n    \"status\": \"Pendente\",\n    \"redirectUrl\": \"https://sandbox.eupago.pt/api/extern/paybylink/form/af3df607c6724870be962a69cac30b99\"\n}"
                  }
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "transactionStatus": {
                      "type": "string",
                      "example": "Success"
                    },
                    "transactionID": {
                      "type": "string",
                      "example": "af3df607c6724870be962a69cac30b99"
                    },
                    "status": {
                      "type": "string",
                      "example": "Pendente"
                    },
                    "redirectUrl": {
                      "type": "string",
                      "example": "https://sandbox.eupago.pt/api/extern/paybylink/form/af3df607c6724870be962a69cac30b99"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "401",
            "content": {
              "application/json": {
                "examples": {
                  "Auth Missing": {
                    "value": "{\n  \"transactionStatus\": \"Rejected\",\n  \"code\": \"APIKEY_MISSING\",\n  \"text\": \"API Key was not available in the request\"\n}"
                  }
                },
                "schema": {
                  "type": "object",
                  "properties": {
                    "transactionStatus": {
                      "type": "string",
                      "example": "Rejected"
                    },
                    "code": {
                      "type": "string",
                      "example": "APIKEY_MISSING"
                    },
                    "text": {
                      "type": "string",
                      "example": "API Key was not available in the request"
                    }
                  }
                }
              }
            }
          }
        },
        "deprecated": false
      }
    },
...
  },
...
}